% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calibration.R
\name{calibration.plot}
\alias{calibration.plot}
\title{E-for-benefit}
\usage{
calibration.plot(
  matched.patients = NULL,
  limits = list(ymin = -1, ymax = 1, xmin = -1, xmax = 1),
  plot.CI = FALSE,
  ...
)
}
\arguments{
\item{matched.patients}{dataframe; dataframe of matched patients, which contains a vector of predicted treatment effect (tau.hat), predicted treatment effect of matched patients (matched.tau.hat), and observed treatment effect (matched.tau.obs) of matched patients}

\item{limits}{list; indicating the x-axis and y-axis limits, e.g. list(ymin=-1, ymax=1, xmin=-1, xmax=1)}

\item{plot.CI}{boolean; TRUE if you want to plot the confidence interval of the calibration plot of predicted versus observed treatment effect of matched patients}

\item{...}{additional arguments for loess function from loess package}
}
\value{
The output of the E.for.benefit function is a "list" with the following components.

matched.patients

a dataframe containing the matched patients.


build.plot

ggplo2 object of the plot.


quantiles

of treatment effect.


se.quantiles

standard error of quantiles of treatment effect.
}
\description{
This function calculates E-for-benefit statistics.  Please note,
this function is only applicable for binary outcomes.
}
\examples{
library(HTEPredictionMetrics)
set.seed(1)
n <- 100
Y <- sample(0:1, n, replace=TRUE)
W <- sample(0:1, n, replace=TRUE)
X <- matrix(rnorm(n), n, 4)
p.0 <- runif(n)
p.1 <- runif(n)
tau.hat <- runif(n)
EB.out <- E.for.Benefit(Y=Y, W=W, X=X, p.0=p.0, p.1=p.1, tau.hat=tau.hat,
                        CI=TRUE, nr.bootstraps=100, message=TRUE,
                        matched.patients=NULL,
                        measure="nearest", distance="mahalanobis",
                        estimand=NULL, replace=FALSE)
limits <- list(ymin=-1, ymax=1, xmin=-1, xmax=1)
calibration.plot(matched.patients=EB.out$matched.patients, limits=limits, plot.CI=TRUE)
}
