% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CforBenefit.R
\name{C.for.Benefit}
\alias{C.for.Benefit}
\title{C-for-benefit}
\usage{
C.for.Benefit(
  Y,
  W,
  X,
  p.0,
  p.1,
  tau.hat,
  CI = FALSE,
  nr.bootstraps = 50,
  message = TRUE,
  measure = "nearest",
  distance = "mahalanobis",
  estimand = "ATT",
  replace = FALSE,
  ...
)
}
\arguments{
\item{Y}{a vector of outcomes}

\item{W}{a vector of treatment assignment; 1 for active treatment; 0 for control}

\item{X}{a matrix of patient characteristics or individualized treatment effect predictions}

\item{p.0}{a vector of outcome probabilities under control}

\item{p.1}{a vector of outcome probabilities under active treatment}

\item{tau.hat}{a vector of individualized treatment effect predictions}

\item{CI}{boolean; TRUE compute confidence interval; default=FALSE do not compute confidence interval (default=FALSE)}

\item{nr.bootstraps}{boolean; number of bootstraps to use for confidence interval computation (default=1)}

\item{message}{boolean; TRUE display computation time message; FALSE do not display message (default=TRUE)}

\item{measure}{measure option of matchit function from MatchIt package (default="nearest")}

\item{distance}{distance option of matchit function from MatchIt package (default="mahalanobis)}

\item{estimand}{default ATT meaning all treated patients get matched with control patient}

\item{replace}{boolean; TRUE if matching with replacement, FALSE if matching without replacement}

\item{...}{additional arguments for matchit function from MatchIt package}
}
\value{
The output of the C.for.benefit function is a "list" with the following components.

matched.patients

a dataframe containing the matched patients.


c.for.benefit

the resulting C-for-benefit value.


lower.CI

the lower bound of the confidence interval (if CI = TRUE).


 upper.CI

the upper bound of the confidence interval (if CI = TRUE).
}
\description{
This function calculates the C-for-benefit, as proposed by
D. van Klaveren et al. (2018), which measures the discriminative ability of
models predicting individualized treatment effect. The C-for-benefit
corresponds to the probability that from two randomly chosen matched patient
pairs with unequal observed treatment effect, the pair with greater observed
treatment effect also has a higher predicted treatment effect. Observed
treatment effect was defined as the difference between outcomes in pairs of
patients matched on patient characteristics (or on individualized treatment
effect predictions). Predicted treatment effect of a matched pair was defined
as the difference between the predicted outcome probability of the untreated
patient minus the predicted outcome probability of the treated patient.
}
\examples{
library(HTEPredictionMetrics)
n <- 100
Y <- sample(0:1, n, replace=TRUE)
W <- sample(0:1, n, replace=TRUE)
X <- matrix(rnorm(n), n, 3)
p.0 <- runif(n)
p.1 <- runif(n)
tau.hat <- runif(n)
CB.out <- C.for.Benefit(Y=Y, W=W, X=X, p.0=p.0, p.1=p.1, tau.hat=tau.hat,
                        CI=TRUE, nr.bootstraps=100, message=TRUE,
                        measure="nearest", distance="mahalanobis",
                        estimand="ATT", replace=FALSE)
CB.out
}
