% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{match.patients}
\alias{match.patients}
\title{Match patients}
\usage{
match.patients(
  Y,
  W,
  X,
  p.0,
  p.1,
  tau.hat,
  measure = "nearest",
  distance = "mahalanobis",
  estimand = NULL,
  ...
)
}
\arguments{
\item{Y}{a vector of binary outcomes; 1 if an (unfavourable) event; 0 if not}

\item{W}{a vector of treatment assignment; 1 for active treatment; 0 for control}

\item{X}{a matrix of patient characteristics or individualized treatment effect predictions, use as.matrix() if necessary, and do not include Y or W in this matrix}

\item{p.0}{a vector of outcome probabilities under control}

\item{p.1}{a vector of outcome probabilities under active treatment}

\item{tau.hat}{a vector of individualized treatment effect predictions}

\item{measure}{measure option of matchit function from MatchIt package (default="nearest")}

\item{distance}{distance option of matchit function from MatchIt package (default="mahalanobis)}

\item{estimand}{default NULL meaning that all patients in the smallest treatment arm get matched with one patient in other treatment arm; estimand can also be "ATT", "ATC" or "ATE", see Details of matchit function of MatchIt for more information}

\item{...}{additional arguments for matchit function from MatchIt package}
}
\value{
The output of the match.patients function is

df.matched.patients

a dataframe containing the matched patients, thus each individual is included in this dataframe


df.matched.pairs

a dataframe containing only the matched pairs, and not the individuals


discarded

a vector of ID's of the patients omitted that are not matched
}
\description{
This function matches patient pairs using the MatchIt package.
Observed treatment effect was defined as the difference between outcomes in
pairs of patients matched on patient characteristics (or on individualized
treatment effect predictions). Predicted treatment effect of a matched pair
was defined as the difference between the predicted outcome probability of
the untreated patient minus the predicted outcome probability of the treated
patient.  Please note, this function is only applicable for binary outcomes.
}
\examples{
library(HTEPredictionMetrics)
set.seed(1)
n <- 100
Y <- sample(0:1, n, replace=TRUE)
W <- sample(0:1, n, replace=TRUE)
X <- matrix(rnorm(n), n, 3)
p.0 <- runif(n)
p.1 <- runif(n)
tau.hat <- runif(n)
matched.patients <- match.patients(Y=Y, W=W, X=X,
                                   p.0=p.0, p.1=p.1, tau.hat=tau.hat,
                                   CI=FALSE, message=TRUE,
                                   measure="nearest", distance="mahalanobis",
                                   estimand=NULL)
matched.patients
}
