% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rubin_combine.R
\name{Rubin.combine}
\alias{Rubin.combine}
\title{val_surv_mi}
\usage{
Rubin.combine(est, se)
}
\arguments{
\item{est}{vector of estimates}

\item{se}{vector of standard errors}
}
\value{
The output of the val_surv_mi function is a "list" with the following components.

est

Combined estimate


se

Combined standard error
}
\description{
This function calculates intercept, slope, and C-index for risk predictions of multiple imputed data set(s).
}
\examples{
library(HTEPredictionMetrics)
m <- 5 # number of imputations
est <- runif(m)
se <- runif(m)
Rubin.combine(est, se)
}
