% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_surv_mi.R
\name{val.surv.mi}
\alias{val.surv.mi}
\title{val_surv_mi}
\usage{
val.surv.mi(
  p,
  y,
  g = 5,
  main = "",
  save_plots = FALSE,
  time = NULL,
  lim = c(0, 1),
  dist = TRUE
)
}
\arguments{
\item{p}{Matrix with predicted probabilities for imputation i in columns (complete case analysis: one column)}

\item{y}{Time to event outcome as Surv object (time,status)}

\item{g}{Number of risk groups; default=5}

\item{main}{Plot label, default=""}

\item{save_plots}{Save plots to file, default=FALSE}

\item{time}{Time point at which to evaluate the predicted probabilities, default=NULL (not entered), the maximum time point will be taken}

\item{lim}{limit, default=NULL}

\item{dist}{distribution, default=TRUE}
}
\value{
The output of the val_surv_mi function is a "list" with the following components.

main

Plot label


n

number of observations


quants

quantiles


p.mi

predicted probabilities for each imputed data set.


obs.mi

observed probabilities lower bound of 95% confidence interval.


obs.mi.lower

observed probabilities lower bound of 95% confidence interval.


int

intercept


int.lower

intercept lower bound of 95% confidence interval.


int.upper

intercept upper bound of 95% confidence interval.


slope

slope estimate


slope.lower

slope estimate lower bound of 95% confidence terval.


slope.upper

slope estimate upper bound of 95% confidence terval.

cindex

C-index

cindex.lower

C-index lower bound of 95% confidence terval.

cindex.upper

C-index upper bound of 95% confidence terval.
}
\description{
This function calculates intercept, slope, and C-index for risk predictions of multiple imputed data set(s).
}
\examples{
library(HTEPredictionMetrics)
library(survival)
n <- 100
m <- 5 # number of imputations
p <- matrix(runif(n*m, 0, 1), n, m)
time_until_event <- runif(n, 0, 5)
status <- rbinom(n, 1, 0.5)
y <- Surv(time_until_event, status)
g <- 4
main <- "Plot label"
save_plots <- FALSE
time <- 30
val.surv.mi(p=p,y=y,g=g,main=main,save_plots=save_plots,time=time)
}
